chore: add configuration for Cucumber tests in CucumberTestConfig

+++ backend/src/test/java/com/farma_ya/unit/service/OrderServiceTest.java
@@ -50,14 +50,14 @@ class OrderServiceTest {
 
     @BeforeEach
     void setUp() {
-        testUser = new User(1L, "testuser", "test@example.com", "password", Role.USER);
+        testUser = new User(1, "testuser", "test@example.com", "password", Role.USER);
 
         testCart = new Cart();
-        testCart.setId(1L);
+        testCart.setId(1);
         testCart.setUser(testUser);
 
         testOrder = new Order();
-        testOrder.setId(1L);
+        testOrder.setId(1);
         testOrder.setUser(testUser);
         testOrder.setStatus(OrderStatus.PENDIENTE);
         testOrder.setTotalAmount(BigDecimal.valueOf(100.0));
@@ -66,24 +66,24 @@ class OrderServiceTest {
     @Test
     void getOrderById_ExistingOrder_ReturnsOrder() {
         // Given
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(testOrder));
+        when(orderRepository.findById(1)).thenReturn(Optional.of(testOrder));
 
         // When
-        Order result = orderService.getOrderById(1L);
+        Order result = orderService.getOrderById(1);
 
         // Then
         assertThat(result).isNotNull();
-        assertThat(result.getId()).isEqualTo(1L);
-        verify(orderRepository).findById(1L);
+        assertThat(result.getId()).isEqualTo(1);
+        verify(orderRepository).findById(1);
     }
 
     @Test
     void getOrderById_NonExistingOrder_ThrowsException() {
         // Given
-        when(orderRepository.findById(999L)).thenReturn(Optional.empty());
+        when(orderRepository.findById(999)).thenReturn(Optional.empty());
 
         // When & Then
-        assertThatThrownBy(() -> orderService.getOrderById(999L))
+        assertThatThrownBy(() -> orderService.getOrderById(999))
                 .isInstanceOf(ResourceNotFoundException.class)
                 .hasMessageContaining("Orden no encontrada con ID: 999");
     }
@@ -91,11 +91,11 @@ class OrderServiceTest {
     @Test
     void updateOrderStatus_ValidStatus_UpdatesSuccessfully() {
         // Given
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(testOrder));
+        when(orderRepository.findById(1)).thenReturn(Optional.of(testOrder));
         when(orderRepository.save(any(Order.class))).thenReturn(testOrder);
 
         // When
-        Order result = orderService.updateOrderStatus(1L, "PROCESANDO");
+        Order result = orderService.updateOrderStatus(1, "PROCESANDO");
 
         // Then
         assertThat(result).isNotNull();
@@ -106,10 +106,10 @@ class OrderServiceTest {
     @Test
     void updateOrderStatus_InvalidStatus_ThrowsException() {
         // Given
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(testOrder));
+        when(orderRepository.findById(1)).thenReturn(Optional.of(testOrder));
 
         // When & Then
-        assertThatThrownBy(() -> orderService.updateOrderStatus(1L, "INVALID_STATUS"))
+        assertThatThrownBy(() -> orderService.updateOrderStatus(1, "INVALID_STATUS"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("Estado de orden inválido");
     }
@@ -117,11 +117,11 @@ class OrderServiceTest {
     @Test
     void updateOrderStatus_FrontendStatusConversion_ConvertsCorrectly() {
         // Given
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(testOrder));
+        when(orderRepository.findById(1)).thenReturn(Optional.of(testOrder));
         when(orderRepository.save(any(Order.class))).thenReturn(testOrder);
 
         // When - Simular envío desde frontend con "EN_PROCESO"
-        Order result = orderService.updateOrderStatus(1L, "EN_PROCESO");
+        Order result = orderService.updateOrderStatus(1, "EN_PROCESO");
 
         // Then - Debe convertirse a PROCESANDO
         assertThat(result.getStatus()).isEqualTo(OrderStatus.PROCESANDO);
@@ -131,7 +131,7 @@ class OrderServiceTest {
     @Test
     void getOrdersByRepartidor_ReturnsOrdersForDeliveryUser() {
         // Given
-        User repartidor = new User(2L, "repartidor", "rep@example.com", "pass", Role.DELIVERY);
+        User repartidor = new User(2, "repartidor", "rep@example.com", "pass", Role.DELIVERY);
         List<Order> expectedOrders = Arrays.asList(testOrder);
 
         when(orderRepository.findByRepartidor(repartidor)).thenReturn(expectedOrders);
@@ -142,18 +142,18 @@ class OrderServiceTest {
         // Then
         assertThat(result).isNotNull();
         assertThat(result).hasSize(1);
-        assertThat(result.get(0).getId()).isEqualTo(1L);
+        assertThat(result.get(0).getId()).isEqualTo(1);
         verify(orderRepository).findByRepartidor(repartidor);
     }
 
     @Test
     void getDeliveryStats_CalculatesCorrectStats() {
         // Given
-        User repartidor = new User(2L, "repartidor", "rep@example.com", "pass", Role.DELIVERY);
+        User repartidor = new User(2, "repartidor", "rep@example.com", "pass", Role.DELIVERY);
 
-        Order order1 = createTestOrder(1L, OrderStatus.PENDIENTE, BigDecimal.valueOf(50.0));
-        Order order2 = createTestOrder(2L, OrderStatus.PROCESANDO, BigDecimal.valueOf(75.0));
-        Order order3 = createTestOrder(3L, OrderStatus.ENTREGADO, BigDecimal.valueOf(100.0));
+        Order order1 = createTestOrder(1, OrderStatus.PENDIENTE, BigDecimal.valueOf(50.0));
+        Order order2 = createTestOrder(2, OrderStatus.PROCESANDO, BigDecimal.valueOf(75.0));
+        Order order3 = createTestOrder(3, OrderStatus.ENTREGADO, BigDecimal.valueOf(100.0));
 
         List<Order> orders = Arrays.asList(order1, order2, order3);
         when(orderRepository.findByRepartidor(repartidor)).thenReturn(orders);
@@ -172,12 +172,12 @@ class OrderServiceTest {
     @Test
     void assignRepartidor_ValidAssignment_AssignsSuccessfully() {
         // Given
-        User repartidor = new User(2L, "repartidor", "rep@example.com", "pass", Role.DELIVERY);
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(testOrder));
+        User repartidor = new User(2, "repartidor", "rep@example.com", "pass", Role.DELIVERY);
+        when(orderRepository.findById(1)).thenReturn(Optional.of(testOrder));
         when(orderRepository.save(any(Order.class))).thenReturn(testOrder);
 
         // When
-        Order result = orderService.assignRepartidor(1L, repartidor);
+        Order result = orderService.assignRepartidor(1, repartidor);
 
         // Then
         assertThat(result).isNotNull();
@@ -209,11 +209,11 @@ class OrderServiceTest {
         long count = orderService.countOrdersByStatus(OrderStatus.PENDIENTE);
 
         // Then
-        assertThat(count).isEqualTo(5L);
+        assertThat(count).isEqualTo(5);
         verify(orderRepository).countByStatus(OrderStatus.PENDIENTE);
     }
 
-    private Order createTestOrder(Long id, OrderStatus status, BigDecimal total) {
+    private Order createTestOrder(Integer id, OrderStatus status, BigDecimal total) {
         Order order = new Order();
         order.setId(id);
         order.setStatus(status);
